+++ src/main/java/racingcar/repository/RacingGameRepository.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
